Unable to parse issue that has a custom field of type Float.
Fixes https://bitbucket.org/farmas/atlassian.net-sdk/issues/128.

+++ Atlassian.Jira/Remote/CustomFieldValueSerializers.cs
@@ -56,7 +56,7 @@ namespace Atlassian.Jira.Remote
 
         public JToken ToJson(string[] values)
         {
-            return long.Parse(values[0]);
+            return float.Parse(values[0]);
         }
     }
